unveil/requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
async-timeout==4.0.3
cffi==1.17.1
cryptography==43.0.1
Django==5.1.1
django-cors-headers==4.4.0
django-ninja==1.3.0
environs==11.0.0
gunicorn==23.0.0
pyjwt>=2.9.0
marshmallow==3.22.0
packaging==24.1
pillow==10.4.0
platformshconfig==2.4.0
psycopg2-binary==2.9.9
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
redis==5.0.8
sqlparse==0.5.1
typing_extensions==4.12.2
